+++ test/ref_inspector/database/reload_test.exs
@@ -1,16 +1,18 @@
 defmodule RefInspector.Database.ReloadTest do
   use ExUnit.Case, async: false
 
+  import ExUnit.CaptureLog
+
   @fixture_search "referers_search.yml"
   @fixture_social "referers_social.yml"
 
   setup do
     app_files = Application.get_env(:ref_inspector, :database_files)
-
-    Application.put_env(:ref_inspector, :database_files, [])
+    app_path = Application.get_env(:ref_inspector, :database_path)
 
     on_exit(fn ->
       Application.put_env(:ref_inspector, :database_files, app_files)
+      Application.put_env(:ref_inspector, :database_path, app_path)
     end)
   end
 
@@ -28,5 +30,31 @@ defmodule RefInspector.Database.ReloadTest do
 
     assert RefInspector.parse("http://www.google.com/test").source == :unknown
     assert RefInspector.parse("http://twitter.com/test").source == "Twitter"
+  end
+
+  test "warns about missing files configuration" do
+    Application.put_env(:ref_inspector, :database_files, [])
+
+    log =
+      capture_log(fn ->
+        RefInspector.reload()
+        :timer.sleep(100)
+      end)
+
+    assert log =~ ~r/no database files/i
+    assert [] == RefInspector.Database.list()
+  end
+
+  test "warns about missing path configuration" do
+    Application.delete_env(:ref_inspector, :database_path)
+
+    log =
+      capture_log(fn ->
+        RefInspector.reload()
+        :timer.sleep(100)
+      end)
+
+    assert log =~ ~r/no database path.*/i
+    assert [] == RefInspector.Database.list()
   end
 end